test: separate multi tenancy test cases
There's a clear distinction between tests that require MT to be enabled, and once that require it to be disabled. It's
easiest to split this test in 2 separate classes.
